requirements.txt
streamlit==1.39.0
PyPDF2==3.0.1
python-docx==1.1.2
fpdf==1.7.2
textblob==0.18.0
pyrebase4==4.8.0
google-generativeai==0.7.2
setuptools
scikit-learn==1.5.2
